# psd-render-api/openapi-current.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PSD 渲染服务平台 API",
    "description": "## 概述\nPSD 模板渲染服务平台对外提供 PSD 模板管理、资产上传、渲染任务提交与查询的 HTTP API。\n\n## 鉴权\n所有 `/v1/*` 接口需在 `Authorization` 头携带 API Key（格式：`Bearer sk_live_xxx`）。\nAPI Key 通过 Admin 控制台创建，每个 Key 可配置：\n- 限流（rateLimitPerMin）：每分钟最大请求数\n- 日配额（quotaPerDay）：每天最大请求数\n- 作用域（scopes）：限定可调用的接口\n- IP 白名单（ipWhitelist）\n- 默认 Webhook URL / Secret\n\n## 幂等性\n`POST /v1/render-jobs` 支持通过 `Idempotency-Key` 头实现幂等：\n同租户同 key 重复提交将返回相同任务，不重复创建。\n\n## Webhook\n任务终态（succeeded / failed / cancelled）时自动投递 Webhook 到提交时声明的 `webhookUrl`。\n投递包含 HMAC-SHA256 签名头 `X-Render-Signature`，可校验来源。\n失败自动重试，指数退避：1m / 5m / 15m / 60m / 240m / 960m。\n\n## 错误响应\n所有错误响应统一为 `{ \"error\": \"<CODE>\", \"message\": \"<描述>\", \"details\"?: [...] }`。",
    "version": "0.3.0",
    "contact": {
      "name": "PSD 渲染服务平台"
    }
  },
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "http",
        "scheme": "bearer",
        "description": "API Key（格式：sk_live_xxx）",
        "bearerFormat": "sk_live_xxx"
      },
      "adminSession": {
        "type": "http",
        "scheme": "bearer",
        "description": "Admin Session Token（登录后获得）",
        "bearerFormat": "session_token"
      },
      "workerToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "Worker 访问令牌（注册后获得）",
        "bearerFormat": "wkr_xxx"
      },
      "storageSignedToken": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "存储签名令牌（格式：Bearer <token>，由 generateUploadUrl/generateDownloadUrl 签发，仅 local 模式可用。也支持 ?token= 查询参数，但不推荐）"
      },
      "workerRegisterSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "Worker 注册凭据（格式：Bearer <WORKER_REGISTER_SECRET> 或一次性配对码/token，也可通过 X-Worker-Register-Secret 头传递）"
      }
    },
    "schemas": {
      "def-0": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "错误码"
          },
          "message": {
            "type": "string",
            "description": "错误描述"
          },
          "details": {
            "type": "array",
            "description": "可选，详细错误信息"
          }
        },
        "required": [
          "error",
          "message"
        ],
        "title": "ErrorResponse"
      }
    }
  },
  "paths": {
...
    "/v1/templates/upload-url": {
      "post": {
        "summary": "获取 PSD 文件上传地址",
        "tags": [
          "templates"
        ],
        "description": "获取 PSD 文件预签名上传地址。PSD 文件将作为模板版本底稿，触发解析后生成图层树。",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "fileName"
                ],
                "properties": {
                  "fileName": {
                    "type": "string",
                    "minLength": 1,
                    "description": "PSD 文件名（含 .psd 扩展名）"
                  },
                  "mimeType": {
                    "type": "string",
                    "description": "可选，默认 image/vnd.adobe.photoshop"
                  }
                }
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "uploadUrl",
                    "method",
                    "headers",
                    "objectKey",
                    "expiresAt"
                  ],
                  "properties": {
                    "uploadUrl": {
                      "type": "string"
                    },
                    "method": {
                      "type": "string"
                    },
                    "headers": {
                      "type": "object",
                      "additionalProperties": true
                    },
                    "objectKey": {
                      "type": "string"
                    },
                    "expiresAt": {
                      "type": "string",
                      "format": "date-time"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "401": {
            "description": "API Key 无效或缺失",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "API Key 无效或缺失"
                }
              }
            }
          },
          "403": {
            "description": "IP 白名单拒绝 / 作用域不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "IP 白名单拒绝 / 作用域不足"
                }
              }
            }
          },
          "429": {
            "description": "触发限流或日配额耗尽",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "触发限流或日配额耗尽"
                }
              }
            }
          }
        }
      }
    },
    "/v1/templates": {
      "post": {
        "summary": "触发 PSD 解析并创建模板",
        "tags": [
          "templates"
        ],
        "description": "基于已上传的 PSD 文件 objectKey 触发解析，生成模板与首个版本（DRAFT 状态），返回模板 ID 与图层树。",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "objectKey",
                  "name"
                ],
                "properties": {
                  "objectKey": {
                    "type": "string",
                    "minLength": 1,
                    "description": "上传 PSD 后获得的 objectKey"
                  },
                  "name": {
                    "type": "string",
                    "minLength": 1,
                    "description": "模板名称"
                  },
                  "psMinVersion": {
                    "type": "string",
                    "description": "可选，最低 PS 版本要求，默认 25.0"
                  }
                }
              }
            }
          },
          "required": true
        },
        "security": [
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "创建成功，返回模板 ID 与解析得到的图层树",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "创建成功，返回模板 ID 与解析得到的图层树",
                  "properties": {
                    "templateId": {
                      "type": "string",
                      "description": "模板 ID（tpl_xxx）"
                    },
                    "version": {
                      "type": "integer",
                      "description": "版本号（首个版本为 1）"
                    },
                    "layerSchema": {
                      "type": "object",
                      "additionalProperties": true,
                      "description": "解析得到的图层树结构"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "401": {
            "description": "API Key 无效或缺少 tenantId",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "API Key 无效或缺少 tenantId"
                }
              }
            }
          },
          "403": {
            "description": "IP 白名单拒绝 / 作用域不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "IP 白名单拒绝 / 作用域不足"
                }
              }
            }
          },
          "422": {
            "description": "PSD 解析失败（非 PSD 文件 / 损坏 / 无图层）",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "PSD 解析失败（非 PSD 文件 / 损坏 / 无图层）"
                }
              }
            }
          },
          "429": {
            "description": "触发限流或日配额耗尽",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "触发限流或日配额耗尽"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "模板列表",
        "tags": [
          "templates"
        ],
        "description": "获取所有模板（含状态、版本数量）。",
        "security": [
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "templates"
                  ],
                  "properties": {
                    "templates": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "API Key 无效或缺少 tenantId",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "API Key 无效或缺少 tenantId"
                }
              }
            }
          },
          "403": {
            "description": "IP 白名单拒绝 / 作用域不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "IP 白名单拒绝 / 作用域不足"
                }
              }
            }
          },
          "429": {
            "description": "触发限流或日配额耗尽",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "触发限流或日配额耗尽"
                }
              }
            }
          }
        }
      }
    },
    "/v1/templates/{id}/layer-bindings": {
      "put": {
        "summary": "保存图层绑定配置",
        "tags": [
          "templates"
        ],
        "description": "为模板的每个可替换图层配置绑定（bindingId、类型、是否必填、接受的格式等）。",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "bindings"
                ],
                "properties": {
                  "bindings": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "required": [
                        "bindingId",
                        "layerId",
                        "layerPath",
                        "type"
                      ],
                      "properties": {
                        "bindingId": {
                          "type": "string",
                          "description": "业务侧自定义的绑定 ID（如图层别名）"
                        },
                        "layerId": {
                          "type": "integer",
                          "description": "PSD 图层 ID"
                        },
                        "layerPath": {
                          "type": "string",
                          "description": "PSD 图层路径"
                        },
                        "type": {
                          "type": "string",
                          "enum": [
                            "smartObject",
                            "text",
                            "pixel"
                          ],
                          "description": "绑定类型允许智能对象、文字和像素图层"
                        },
                        "required": {
                          "type": "boolean",
                          "default": true
                        },
                        "label": {
                          "type": "string"
                        },
                        "acceptedFormats": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        },
                        "fit": {
                          "type": "string",
                          "enum": [
                            "cover",
                            "contain",
                            "stretch"
                          ]
                        },
                        "maxLength": {
                          "type": "integer",
                          "minimum": 1
                        },
                        "defaultFontVersionId": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "required": true
        },
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "模板 ID（tpl_xxx）"
          }
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok",
                    "templateId"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "description": "操作是否成功"
                    },
                    "templateId": {
                      "type": "string",
                      "description": "模板 ID（tpl_xxx）"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "401": {
            "description": "API Key 无效或缺少 tenantId",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "API Key 无效或缺少 tenantId"
                }
              }
            }
          },
          "403": {
            "description": "IP 白名单拒绝 / 作用域不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "IP 白名单拒绝 / 作用域不足"
                }
              }
            }
          },
          "404": {
            "description": "模板不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "模板不存在"
                }
              }
            }
          },
          "429": {
            "description": "触发限流或日配额耗尽",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "触发限流或日配额耗尽"
                }
              }
            }
          }
        }
      }
    },
    "/v1/templates/{id}/publish": {
      "post": {
        "summary": "发布模板版本",
        "tags": [
          "templates"
        ],
        "description": "将模板当前版本从 DRAFT 切换为 PUBLISHED，发布后才能提交渲染任务。",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "模板 ID（tpl_xxx）"
          }
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "ok",
                    "templateId"
                  ],
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "description": "操作是否成功"
                    },
                    "templateId": {
                      "type": "string",
                      "description": "模板 ID（tpl_xxx）"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Default Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "401": {
            "description": "API Key 无效或缺少 tenantId",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "API Key 无效或缺少 tenantId"
                }
              }
            }
          },
          "403": {
            "description": "IP 白名单拒绝 / 作用域不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "IP 白名单拒绝 / 作用域不足"
                }
              }
            }
          },
          "404": {
            "description": "模板不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "模板不存在"
                }
              }
            }
          },
          "409": {
            "description": "模板版本非 DRAFT 状态，无法发布",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "模板版本非 DRAFT 状态，无法发布"
                }
              }
            }
          },
          "429": {
            "description": "触发限流或日配额耗尽",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "触发限流或日配额耗尽"
                }
              }
            }
          }
        }
      }
    },
    "/v1/templates/{id}": {
      "get": {
        "summary": "模板详情",
        "tags": [
          "templates"
        ],
        "description": "获取模板详情，含图层树、绑定配置、版本列表。",
        "parameters": [
          {
            "schema": {
              "type": "string"
            },
            "in": "path",
            "name": "id",
            "required": true,
            "description": "模板 ID（tpl_xxx）"
          }
        ],
        "security": [
          {
            "apiKey": []
          }
        ],
        "responses": {
          "200": {
            "description": "模板详情，含图层树、绑定配置、版本列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "模板详情，含图层树、绑定配置、版本列表",
                  "properties": {
                    "templateId": {
                      "type": "string"
                    },
                    "code": {
                      "type": "string"
                    },
                    "name": {
                      "type": "string"
                    },
                    "status": {
                      "type": "string",
                      "description": "DRAFT | PUBLISHED | ARCHIVED | DELETED"
                    },
                    "versions": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    },
                    "layerSchema": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "API Key 无效或缺少 tenantId",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "API Key 无效或缺少 tenantId"
                }
              }
            }
          },
          "403": {
            "description": "IP 白名单拒绝 / 作用域不足",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "IP 白名单拒绝 / 作用域不足"
                }
              }
            }
          },
          "404": {
            "description": "模板不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "模板不存在"
                }
              }
            }
          },
          "429": {
            "description": "触发限流或日配额耗尽",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/def-0",
                  "description": "触发限流或日配额耗尽"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}